+++ Migrations/Code/LegacyClassesForIde.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace {
+
+    die('Access denied');
+}
+
+namespace {
+
+    /**
+     * @deprecated since 5.7, will be removed in 8.x
+     */
+    interface Tx_Phpunit_Interface_AccessibleObject extends \OliverKlee\Phpunit\Interfaces\AccessibleObject
+    {
+    }
+
+    /**
+     * @deprecated since 5.7, will be removed in 8.x
+     */
+    class Tx_Phpunit_TestCase extends \OliverKlee\PhpUnit\TestCase
+    {
+    }
+}