+++ TestExtensions/aaa/Tests/Unit/OneTest.php
@@ -7,6 +7,6 @@
  *
  * @author Nicole Cordes <nicole.cordes@googlemail.com>
  */
-class Tx_Aaa_Tests_OneTest extends \Tx_Phpunit_TestCase
+class Tx_Aaa_Tests_OneTest extends \OliverKlee\PhpUnit\TestCase
 {
 }