+++ TestExtensions/aaa/Tests/Unit/fixtures/AnotherTest.php
@@ -7,6 +7,6 @@
  *
  * @author Nicole Cordes <nicole.cordes@googlemail.com>
  */
-class Tx_Aaa_Tests_Fixtures_AnotherTest extends \Tx_Phpunit_TestCase
+class Tx_Aaa_Tests_Fixtures_AnotherTest extends \OliverKlee\PhpUnit\TestCase
 {
 }