+++ Tests/Unit/Service/TestFinderTest.php
@@ -2,6 +2,7 @@
 
 namespace OliverKlee\Phpunit\Tests\Unit\Service;
 
+use OliverKlee\PhpUnit\TestCase;
 use TYPO3\CMS\Core\SingletonInterface;
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\PathUtility;
@@ -12,7 +13,7 @@ use TYPO3\CMS\Core\Utility\PathUtility;
  *
  * @author Oliver Klee <typo3-coding@oliverklee.de>
  */
-class TestFinderTest extends \Tx_Phpunit_TestCase
+class TestFinderTest extends TestCase
 {
     /**
      * @var \Tx_Phpunit_Service_TestFinder