+++ Tests/Unit/ViewHelpers/ExtensionSelectorViewHelperTest.php
@@ -2,6 +2,7 @@
 
 namespace OliverKlee\Phpunit\Tests\Unit\ViewHelpers;
 
+use OliverKlee\PhpUnit\TestCase;
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\PathUtility;
 use TYPO3\CMS\Core\Utility\VersionNumberUtility;
@@ -12,7 +13,7 @@ use TYPO3\CMS\Lang\LanguageService;
  *
  * @author Nicole Cordes <nicole.cordes@googlemail.com>
  */
-class ExtensionSelectorViewHelperTest extends \Tx_Phpunit_TestCase
+class ExtensionSelectorViewHelperTest extends TestCase
 {
     /**
      * @var \Tx_Phpunit_ViewHelpers_ExtensionSelectorViewHelper