+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Auto-release to the TER (#67)
 
 ### Changed
+- Namespace the non-deprecated classes and interfaces (#96)
 - Update PHPUnit to 5.7 (#93, #94)
 - Stop using a PHAR for including libraries (#87)
 - Use spaces for indenting SQL and .htaccess files (#78)